Deployment/scripts/ai_hub_scripts/requirements.txt
azure-mgmt-resource==23.1.0b2
msal==1.24.0b1
azure-ai-resources